+++ tests/test.lunaris.index.js
@@ -123,6 +123,11 @@ describe('lunaris index', () => {
     should(index.setPagination).be.a.Function();
   });
 
+  it('createUrl() should be defined', () => {
+    should(index.createUrl).be.ok();
+    should(index.createUrl).be.a.Function();
+  });
+
   it('begin() should be defined', () => {
     should(index.begin).be.ok();
     should(index.begin).be.a.Function();